refactor: adapterを移動
