+++ app/src/main/java/com/example/snap_develop/activity/DisplayCommentActivity.java
@@ -17,10 +17,10 @@ import androidx.lifecycle.ViewModelProvider;
 import com.example.snap_develop.MainApplication;
 import com.example.snap_develop.MyDebugTree;
 import com.example.snap_develop.R;
-import com.example.snap_develop.adapter.DisplayCommentAdapter;
 import com.example.snap_develop.bean.PostBean;
 import com.example.snap_develop.bean.UserBean;
 import com.example.snap_develop.databinding.ActivityDisplayCommentBinding;
+import com.example.snap_develop.view.adapter.DisplayCommentAdapter;
 import com.example.snap_develop.viewModel.FollowViewModel;
 import com.example.snap_develop.viewModel.PostViewModel;
 import com.example.snap_develop.viewModel.UserViewModel;
