+++ app/src/main/java/com/example/snap_develop/activity/FollowingListActivity.java
@@ -13,9 +13,9 @@ import androidx.lifecycle.ViewModelProvider;
 
 import com.example.snap_develop.MyDebugTree;
 import com.example.snap_develop.R;
-import com.example.snap_develop.adapter.FollowListAdapter;
 import com.example.snap_develop.bean.UserBean;
 import com.example.snap_develop.databinding.ActivityFollowingListBinding;
+import com.example.snap_develop.view.adapter.FollowListAdapter;
 import com.example.snap_develop.viewModel.FollowViewModel;
 import com.example.snap_develop.viewModel.UserViewModel;
 
@@ -84,7 +84,7 @@ public class FollowingListActivity extends AppCompatActivity implements View.OnC
         }
     }
 
-    
+
     @Override
     public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
         UserBean userBean = mFollowList.get(position);
