+++ app/src/main/java/com/example/snap_develop/activity/PostSearchActivity.java
@@ -18,10 +18,10 @@ import androidx.lifecycle.ViewModelProvider;
 
 import com.example.snap_develop.MyDebugTree;
 import com.example.snap_develop.R;
-import com.example.snap_develop.adapter.PostSearchAdapter;
 import com.example.snap_develop.bean.PostBean;
 import com.example.snap_develop.bean.UserBean;
 import com.example.snap_develop.databinding.ActivityPostSearchBinding;
+import com.example.snap_develop.view.adapter.PostSearchAdapter;
 import com.example.snap_develop.viewModel.PostViewModel;
 import com.example.snap_develop.viewModel.UserViewModel;
